+++ src/app/components/index.ts
@@ -1,7 +1,7 @@
 // Re-export all components from their respective folders
 export * from './layout';
 export * from './ui';
-export * from './topic';
+export * from './topics-ui';
 export * from './progress';
 export * from './questions';
 export * from './marketing';
