+++ src/app/components/questions/index.ts
@@ -1,5 +1,4 @@
 export * from './QuestionList';
 export * from './QuestionWithAnswer';
 export * from './BookmarkButton';
-export { default as ActivityTimeGrid } from './ActivityTimeGrid';
 export * from './ResourceList';
